Enforce warning-free Swift builds
